LPS-53627 extract print functionality to its own module
