+++ modules/apps/journal/journal-content-asset-addon-entries/journal-content-print-asset-addon-entry/build.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+<!DOCTYPE project>
+
+<project>
+	<property name="import.shared" value="../journal-content-web" />
+
+	<import file="../../../../../tools/sdk/build-common-osgi-plugin.xml" />
+
+	<property name="auto.deploy.dir" value="${liferay.home}/osgi/modules" />
+</project>
