+++ modules/apps/journal/journal-content-web/src/META-INF/resources/view.jsp
@@ -53,39 +53,6 @@ AssetRendererFactory assetRendererFactory = AssetRendererFactoryRegistryUtil.get
 						<div class="user-tool-asset-addon-entries">
 							<liferay-ui:asset-addon-entry-display assetAddonEntries="<%= journalContentDisplayContext.getSelectedUserToolAssetAddonEntries() %>" />
 
-								<c:if test="<%= journalContentDisplayContext.isEnablePrint() %>">
-									<c:choose>
-										<c:when test="<%= journalContentDisplayContext.isPrint() %>">
-											<aui:script>
-												print();
-											</aui:script>
-										</c:when>
-										<c:otherwise>
-											<portlet:renderURL var="printPageURL" windowState="<%= LiferayWindowState.POP_UP.toString() %>">
-												<portlet:param name="groupId" value="<%= String.valueOf(articleDisplay.getGroupId()) %>" />
-												<portlet:param name="articleId" value="<%= articleDisplay.getArticleId() %>" />
-												<portlet:param name="page" value="<%= String.valueOf(articleDisplay.getCurrentPage()) %>" />
-												<portlet:param name="viewMode" value="<%= Constants.PRINT %>" />
-											</portlet:renderURL>
-
-											<div class="print-action">
-												<liferay-ui:icon
-													iconCssClass="icon-print"
-													label="<%= true %>"
-													message='<%= LanguageUtil.format(request, "print-x-x", new Object[] {"hide-accessible", HtmlUtil.escape(articleDisplay.getTitle())}, false) %>'
-													url='<%= "javascript:" + renderResponse.getNamespace() + "printPage();" %>'
-												/>
-											</div>
-
-											<aui:script>
-												function <portlet:namespace />printPage() {
-													window.open('<%= printPageURL %>', '', 'directories=0,height=480,left=80,location=1,menubar=1,resizable=1,scrollbars=yes,status=0,toolbar=0,top=180,width=640');
-												}
-											</aui:script>
-										</c:otherwise>
-									</c:choose>
-								</c:if>
-
 								<c:if test="<%= journalContentDisplayContext.isEnableConversions() && !journalContentDisplayContext.isPrint() %>">
 									<div class="export-actions">
 										<liferay-ui:icon-list>
