+++ modules/apps/journal/journal-content-web/src/com/liferay/journal/content/web/context/JournalContentDisplayContext.java
@@ -537,17 +537,6 @@ public class JournalContentDisplayContext {
 		return _enableConversions;
 	}
 
-	public boolean isEnablePrint() {
-		if (_enablePrint != null) {
-			return _enablePrint;
-		}
-
-		_enablePrint = GetterUtil.getBoolean(
-			_portletPreferences.getValue("enablePrint", null));
-
-		return _enablePrint;
-	}
-
 	public boolean isEnableRatings() {
 		if (_enableRatings != null) {
 			return _enableRatings;
@@ -777,7 +766,6 @@ public class JournalContentDisplayContext {
 	private Boolean _enableCommentRatings;
 	private Boolean _enableComments;
 	private Boolean _enableConversions;
-	private Boolean _enablePrint;
 	private Boolean _enableRatings;
 	private Boolean _enableRelatedAssets;
 	private Boolean _enableViewCountIncrement;
